Remove all torchtext datasets
Move the remaining code in the CQA-based classes
